Use symlinks and front-matter defaults to avoid duplicating index
We cannot simply have the contents of `index.html` as an include,
because includes do not evaluate liquid expressions in them. The
alternative would be to create an include for each of the bodies inside
the for loops.

Instead, we create a single `index.html` and two symlinks pointing to
it. Then, we make use of jekyll's front-matter defaults to give each of
the symlinks their own language and permalink.